refactor sidecar::AppInstance in to separate file
