+++ sidecar/src/service/telemetry/mod.rs
@@ -1,0 +1,2 @@
+pub use app_instance::AppInstance;
+mod app_instance;
